Initial placement of relays
Still need to do a lot of work. Placement is temporary.

LEDs and the rest of the components need to be placed.

One problem is that I can't move what they originally did. Option #1: Redo what they did. (not happening) Option #2: Try adding their schematic as a part then placing it on a new shm. Option #3: Try to find a way to move it online.

Signed-off-by: Scott Lerner <scottphiliplerner@gmail.com>